% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigestionExperiment.R
\name{buildDigestDf}
\alias{buildDigestDf}
\title{Helper function of simulateCoDigest to build digest dataframe}
\usage{
buildDigestDf(fragments, enzymeName, sequenceLen)
}
\description{
Helper function of simulateCoDigest to build digest dataframe
}
\keyword{internal}
