% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Enzymes}
\alias{Enzymes}
\title{Type 2 Restriction Enzymes Metadata}
\format{
A matrix with 3 columns:
\describe{
\item{Name}{Name of restriction enzyme.}
\item{Prototype}{Prototype of restriction enzyme.}
\item{RecognitionSeq}{Recognition sequence of restriction enzyme, including cleavage site.}
}
}
\source{
Roberts, R.J., Vincze, T., Posfai, J., Macelis, D. REBASE-a database for DNA restriction and modification: enzymes, genes and genomes. Nucleic Acids Res. 43: D298-D299 (2015).
}
\usage{
Enzymes
}
\description{
Dataset of type 2 restriction enzymes from REABSE version 511 database from Oct 28 2025.
}
\examples{
\dontrun{
 Enzymes
 View(Enzymes)
 head(Enzymes)
}
}
\keyword{datasets}
