% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runREDesignR.R
\name{runREDesignR}
\alias{runREDesignR}
\title{Launch Shiny App for REDesignR}
\usage{
runREDesignR()
}
\value{
No return value but opens up a Shiny app.
}
\description{
A function that launches the Shiny app for REDesignR.
The purpose of this app is to \link{...} The code of the app has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

REDesignR::runREDesignR()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
