% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigestionExperiment.R
\name{simulateCoDigest}
\alias{simulateCoDigest}
\title{Simulates multi-restriction enzyme digestion}
\usage{
simulateCoDigest(DNA, enzymes)
}
\arguments{
\item{DNA}{A DNAString representing the DNA sequence to be digested in the simulation.}

\item{enzymes}{A table of 2 restriction enzymes containing the following at minimum:
\itemize {
\item Name - Names of the restriction enzymes.
\item RecognitionSeq - Recognition sequences of the corresponding restriction enzyme, including the cleavage site.
}}
}
\value{
Returns a table of digestion results.
\itemize{
\item Enzyme - Name of the restriction enzyme used for digestion.
\item FragmentID - A numeric value representing the order of resulting digested DNA fragments.
\item Start - The starting position of the DNA fragment, in reference to the pre-digested DNA sequence.
\item End - The ending position of the DNA fragment, in reference to the pre-digested DNA sequence.
\item Length - The length of the resulting DNA fragment after digestion.
}
}
\description{
A function that simulates a co-digestion experiment given a DNA sequence and the recognition sequences of 2 restriction enzymes.
}
\examples{
# Using restriction enzymes available with package.

# Example 1:
# Simulate co-digestion with RE AaaI and AagI.
simulateCoDigest(DNAString("CGGCCGATCGATCGGCCG"), Enzymes[c(1,4),])

# Example 2:
# Simulate digestion whne no clevage sites are found.
simulateCoDigest(DNAString("AGGATAAACAA"), Enzymes[c(1,4),])

}
\references{
Müller K, Wickham H (2025). \emph{tibble: Simple Data Frames}. doi:10.32614/CRAN.package.tibble \url{https://doi.org/10.32614/CRAN.package.tibble}, R package version 3.3.0, \url{https://CRAN.R-project.org/package=tibble}.
Pagès H, Aboyoun P, Gentleman R, DebRoy S (2025). \emph{Biostrings: Efficient manipulation of biological strings}. doi:10.18129/B9.bioc.Biostrings \url{https://doi.org/10.18129/B9.bioc.Biostrings}, R package version 2.78.0, \url{https://bioconductor.org/packages/Biostrings}.
Pagès H, Lawrence M, Aboyoun P (2025). \emph{S4Vectors: Foundation of vector-like and list-like containers in Bioconductor}. doi:10.18129/B9.bioc.S4Vectors \url{https://doi.org/10.18129/B9.bioc.S4Vectors}, R package version 0.48.0, \url{https://bioconductor.org/packages/S4Vectors}.
Wickham H (2019). \emph{assertthat: Easy Pre and Post Assertions}. doi:10.32614/CRAN.package.assertthat \url{https://doi.org/10.32614/CRAN.package.assertthat}, R package version 0.2.1, \url{https://CRAN.R-project.org/package=assertthat}.
Wright ES (2024). “Fast and Flexible Search for Homologous Biological Sequences with DECIPHER v3.” \emph{The R Journal}, \emph{16}(2), 191-200.
}
