% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigestVisualization.R
\name{simulateGel}
\alias{simulateGel}
\title{Simulate an agarose gel of restriction digestion fragments}
\usage{
simulateGel(codigestDf, multiDigest = TRUE, labelFragments = FALSE)
}
\arguments{
\item{codigestDf}{A tibble or data frame of co-digestion results in the
format outputted by \code{\link[=simulateCoDigest]{simulateCoDigest()}}.}

\item{multiDigest}{A boolean indicating whether \code{codigestDf} resulted from a
single-digest (\code{FALSE}) or a co-digest (\code{TRUE}), the latter being the default.}

\item{labelFragments}{A boolean indicating whether plotted fragments should
be labelled with the \emph{FragmentID} (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
A plot representing an agarose gel of the fragments resulting
from \code{\link[=simulateCoDigest]{simulateCoDigest()}}.
Each lane corresponds to one digestion condition (single enzyme or co-digest),
and each band represents a DNA fragment.

Band opacity represents gel band "brightness" which reflects the number of
repeated fragment sizes, mimicking real gel intensity variation.

If \code{labelFragments = TRUE}, each band is annotated with the corresponding
\code{FragmentID}'s.
}
\description{
This function produces a simulated agarose gel diagram from digestion results
generated by \code{\link[=simulateCoDigest]{simulateCoDigest()}}.
}
\examples{
# get sample data
data(Enzymes, package = "REDesignR")
dna <- Biostrings::DNAString("CGGCCGATCGATCGGCCG")

# Example 1: ---------------------------------------------------------------
# Multi-lane gel for a co-digest
co <- simulateCoDigest(dna, Enzymes[c(1, 4), ])
simulateGel(co$digestDf)

# Label fragments
simulateGel(co$digestDf, labelFragments = TRUE)

# Example 2: ---------------------------------------------------------------
# Single-digest gel
single <- simulateCoDigest(dna, Enzymes[c(1, 6), ])
simulateGel(single$digestDf, multiDigest = FALSE)

}
\references{
Wickham H, Pedersen T, Seidel D (2025). \emph{scales: Scale Functions for
Visualization.} R package version 1.4.0, https://scales.r-lib.org.
}
