% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigestVisualization.R
\name{plotFragments}
\alias{plotFragments}
\title{Visualize distribution of digest fragments' characteristic(s)}
\usage{
plotFragments(
  codigestDf,
  seq_length = NULL,
  show_lengths = TRUE,
  gel_style = TRUE
)
}
\description{
TODO description
}
\details{
TODO examples
}
