% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigestVisualization.R
\name{plotRestrictionMap}
\alias{plotRestrictionMap}
\title{Visualize Restriction Enzyme co-digestion results}
\usage{
plotRestrictionMap(codigestDf, showLengths)
}
\arguments{
\item{codigestDf}{A dataframe of codigestion results in the format outputted by \code{\link[=simulateCoDigest]{simulateCoDigest()}}.}

\item{show_lengths}{A boolean dictating whether numerical fragment lengths are shown as overlay on the plot; TRUE by default.}
}
\value{
A Plot of the results of a co-digestion experiment, either as an agarose gel or linear restriction map view and/or overlap of fragment lengths.

TODO examples
}
\description{
TODO description
}
\references{
Pedersen T (2025). \emph{patchwork: The Composer of Plots.} R package version 1.3.2.9000, https://patchwork.data-imaginist.com.
}
\seealso{
\code{\link[=simulateCoDigest]{simulateCoDigest()}}
}
