% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runREDesignR.R
\name{runREDesignR}
\alias{runREDesignR}
\title{Launch Shiny App for REDesignR}
\usage{
runREDesignR()
}
\value{
No return value but opens up a Shiny app.
}
\description{
A function that launches the Shiny app for REDesignR.
The purpose of this app is to simulate restriction enzyme co-digestion and
generate restriction maps and gel simulations to aid in digestion experiment
design optimization. The code of the app has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

REDesignR::runREDesignR()
}

}
\references{
Grolemund, G. (2015). \emph{Learn Shiny - Video Tutorials.}
https://shiny.rstudio.com/tutorial/

\emph{Shiny - File Download}. (2014, July 29). Shiny.
https://shiny.posit.co/r/gallery/widgets/file-download/‌

\emph{Shiny - File Upload.} (2014, July 29). Shiny.
https://shiny.posit.co/r/gallery/widgets/file-upload/‌
}
