% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigestVisualization.R
\name{simulateGel}
\alias{simulateGel}
\title{Agarose Gel Simulation of RE Digests}
\usage{
simulateGel(codigestDf, multiDigest = TRUE, labelFragments = FALSE)
}
\arguments{
\item{codigestDf}{A tibble or data frame of co-digestion results in the format outputted by \code{\link[=simulateCoDigest]{simulateCoDigest()}}.}

\item{multiDigest}{A boolean indicating whether \code{codigestDf} resulted from a single-digest (\code{FALSE}) or a co-digest (\code{TRUE}), the latter being the default.}

\item{labelFragments}{A boolean indicating whether plotted fragments should be labelled with the \emph{FragmentID} (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
A plot representing an agarose gel of the fragments resulting from \code{\link[=simulateCoDigest]{simulateCoDigest()}}.
}
\description{
TODO description
}
\examples{
\dontrun{
TODO
}

}
\references{
Wickham H, Pedersen T, Seidel D (2025). \emph{scales: Scale Functions for Visualization.} R package version 1.4.0, https://scales.r-lib.org.
}
