% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigestVisualization.R
\name{simulateGel}
\alias{simulateGel}
\title{Agarose Gel Simulation of RE Digests}
\usage{
simulateGel(codigestDf, multiDigest = TRUE, labelFragments = FALSE)
}
\arguments{
\item{codigestDf}{A dataframe of codigestion results in the format outputted by \code{\link[=simulateCoDigest]{simulateCoDigest()}}.}

\item{multiDigest}{A boolean indicating whether \code{codigestDf} resulted from a single-digest (\code{FALSE}) or a co-digest (\code{TRUE}), the latter being the default.}

\item{labelFragments}{A boolean indicating whether plotted fragments should be labelled with the \emph{FragmentID} (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
A plot representing an agarose gel of the fragments resulting from \code{\link[=simulateCoDigest]{simulateCoDigest()}}.
}
\description{
TODO description
}
\examples{
\dontrun{
TODO
data(Enzymes, package = "REDesignR")  # load data
# Simulate co-digestion with RE AaaI and AagI.
TODO
coDigestDf <- simulateCoDigest(Biostrings::DNAString("CGGCCGATCGATCGGCCG"), REDesignR::Enzymes[c(1,4),])

# Example 1: ---------------------------------------------------------------
simulateGel(coDigestDf) # by default: multiDigest = TRUE, labelFragments = FALSE

# Example 2: ---------------------------------------------------------------
simulateGel(coDigestDf, labelFragments = TRUE)  # by default: multiDigest = TRUE

# Example 3: ---------------------------------------------------------------
# Simulate single-digestion with TODO
TODO
digestDF <- TODO
simulateGel(digestDF, multiDigest = FALSE)  # by default: labelFragments = FALSE
# > REDesignR::simulateGel(testDigestDF)
# > REDesignR::simulateGel(testDigestDF, multiDigest = TRUE)
# > REDesignR::simulateGel(testDigestDF, labelFragments = TRUE)
# > REDesignR::simulateGel(testDigestDF2)
# > REDesignR::simulateGel(testDigestDF2, multiDigest = FALSE)
# > REDesignR::simulateGel(testDigestDF2)
# > REDesignR::simulateGel(testDigestDF2, multiDigest = FALSE)
# > REDesignR::simulateGel(testDigestDF2, multiDigest = FALSE, TRUE)
}

}
\references{
Wickham H, Pedersen T, Seidel D (2025). \emph{scales: Scale Functions for Visualization.} R package version 1.4.0, https://scales.r-lib.org.
}
